list all problem sessions and session solutions

+++ PSDMWebApi/src/main/java/com/masters/PSDMWebApi/repository/SessionRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface SessionRepository extends JpaRepository<Session, Long> {
-    List<Session> findByProblemId(Integer problemId);
+    List<Session> findByProblemId(Long problemId);
 }
